// src/lib/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      clinical_notes: {
        Row: {
          content: string | null
          created_at: string | null
          date_time: string | null
          id: string
          note_type: string | null
          patient_id: string | null
          provider: string | null
        }
        Insert: {
          content?: string | null
          created_at?: string | null
          date_time?: string | null
          id?: string
          note_type?: string | null
          patient_id?: string | null
          provider?: string | null
        }
        Update: {
          content?: string | null
          created_at?: string | null
          date_time?: string | null
          id?: string
          note_type?: string | null
          patient_id?: string | null
          provider?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "clinical_notes_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
      diagnoses: {
        Row: {
          code: string | null
          created_at: string | null
          date_of_diagnosis: string | null
          description: string | null
          id: string
          patient_id: string | null
          status: string | null
        }
        Insert: {
          code?: string | null
          created_at?: string | null
          date_of_diagnosis?: string | null
          description?: string | null
          id?: string
          patient_id?: string | null
          status?: string | null
        }
        Update: {
          code?: string | null
          created_at?: string | null
          date_of_diagnosis?: string | null
          description?: string | null
          id?: string
          patient_id?: string | null
          status?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "diagnoses_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
      lab_results: {
        Row: {
          created_at: string | null
          date_of_test: string | null
          id: string
          patient_id: string | null
          reference_range: string | null
          result: string | null
          test_name: string | null
          units: string | null
        }
        Insert: {
          created_at?: string | null
          date_of_test?: string | null
          id?: string
          patient_id?: string | null
          reference_range?: string | null
          result?: string | null
          test_name?: string | null
          units?: string | null
        }
        Update: {
          created_at?: string | null
          date_of_test?: string | null
          id?: string
          patient_id?: string | null
          reference_range?: string | null
          result?: string | null
          test_name?: string | null
          units?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "lab_results_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
      medications: {
        Row: {
          created_at: string | null
          dosage: string | null
          end_date: string | null
          frequency: string | null
          id: string
          name: string
          patient_id: string | null
          reason: string | null
          route: string | null
          start_date: string | null
        }
        Insert: {
          created_at?: string | null
          dosage?: string | null
          end_date?: string | null
          frequency?: string | null
          id?: string
          name: string
          patient_id?: string | null
          reason?: string | null
          route?: string | null
          start_date?: string | null
        }
        Update: {
          created_at?: string | null
          dosage?: string | null
          end_date?: string | null
          frequency?: string | null
          id?: string
          name?: string
          patient_id?: string | null
          reason?: string | null
          route?: string | null
          start_date?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "medications_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
      patients: {
        Row: {
          address: string | null
          alcohol_consumption: string | null
          allergies: string[] | null
          created_at: string | null
          date_of_birth: string
          email: string | null
          emergency_contact_name: string | null
          emergency_contact_phone: string | null
          emergency_contact_relationship: string | null
          family_history: string[] | null
          gender: string | null
          id: string
          insurance_policy_number: string | null
          insurance_provider: string | null
          name: string
          organization_id: string
          phone: string | null
          registration_date: string | null
          smoking_status: string | null
          status: string | null
          updated_at: string | null
        }
        Insert: {
          address?: string | null
          alcohol_consumption?: string | null
          allergies?: string[] | null
          created_at?: string | null
          date_of_birth: string
          email?: string | null
          emergency_contact_name?: string | null
          emergency_contact_phone?: string | null
          emergency_contact_relationship?: string | null
          family_history?: string[] | null
          gender?: string | null
          id: string
          insurance_policy_number?: string | null
          insurance_provider?: string | null
          name: string
          organization_id: string
          phone?: string | null
          registration_date?: string | null
          smoking_status?: string | null
          status?: string | null
          updated_at?: string | null
        }
        Update: {
          address?: string | null
          alcohol_consumption?: string | null
          allergies?: string[] | null
          created_at?: string | null
          date_of_birth?: string
          email?: string | null
          emergency_contact_name?: string | null
          emergency_contact_phone?: string | null
          emergency_contact_relationship?: string | null
          family_history?: string[] | null
          gender?: string | null
          id?: string
          insurance_policy_number?: string | null
          insurance_provider?: string | null
          name?: string
          organization_id?: string
          phone?: string | null
          registration_date?: string | null
          smoking_status?: string | null
          status?: string | null
          updated_at?: string | null
        }
        Relationships: []
      }
      procedures: {
        Row: {
          code: string | null
          created_at: string | null
          date_of_procedure: string | null
          description: string | null
          id: string
          patient_id: string | null
          provider: string | null
        }
        Insert: {
          code?: string | null
          created_at?: string | null
          date_of_procedure?: string | null
          description?: string | null
          id?: string
          patient_id?: string | null
          provider?: string | null
        }
        Update: {
          code?: string | null
          created_at?: string | null
          date_of_procedure?: string | null
          description?: string | null
          id?: string
          patient_id?: string | null
          provider?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "procedures_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
      vital_signs: {
        Row: {
          blood_pressure_diastolic: number | null
          blood_pressure_systolic: number | null
          created_at: string | null
          date_time: string | null
          heart_rate: number | null
          height: number | null
          id: string
          patient_id: string | null
          respiratory_rate: number | null
          temperature: number | null
          weight: number | null
        }
        Insert: {
          blood_pressure_diastolic?: number | null
          blood_pressure_systolic?: number | null
          created_at?: string | null
          date_time?: string | null
          heart_rate?: number | null
          height?: number | null
          id?: string
          patient_id?: string | null
          respiratory_rate?: number | null
          temperature?: number | null
          weight?: number | null
        }
        Update: {
          blood_pressure_diastolic?: number | null
          blood_pressure_systolic?: number | null
          created_at?: string | null
          date_time?: string | null
          heart_rate?: number | null
          height?: number | null
          id?: string
          patient_id?: string | null
          respiratory_rate?: number | null
          temperature?: number | null
          weight?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "vital_signs_patient_id_fkey"
            columns: ["patient_id"]
            isOneToOne: false
            referencedRelation: "patients"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type TablesInsert<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Insert: infer I
    }
    ? I
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Insert: infer I
      }
      ? I
      : never
    : never

export type TablesUpdate<
  DefaultSchemaTableNameOrOptions extends
    | keyof DefaultSchema["Tables"]
    | { schema: keyof DatabaseWithoutInternals },
  TableName extends DefaultSchemaTableNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"]
    : never = never,
> = DefaultSchemaTableNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaTableNameOrOptions["schema"]]["Tables"][TableName] extends {
      Update: infer U
    }
    ? U
    : never
  : DefaultSchemaTableNameOrOptions extends keyof DefaultSchema["Tables"]
    ? DefaultSchema["Tables"][DefaultSchemaTableNameOrOptions] extends {
        Update: infer U
      }
      ? U
      : never
    : never